To table variants added field image

+++ Inventory/Shared/Variant.razor.cs
@@ -13,8 +13,10 @@ namespace Inventory.Shared
         [Parameter] public ProductEntity Product { get; set; }
         [Inject] private IVariantRepository VariantRepository { get; set; }
         [Inject] private IProductRepository ProductRepository { get; set; }
+        [Inject] private IImageRepository ImageRepository { get; set; }
         [Inject] private ILogger<Login> Logger { get; set; }
         [Inject] private IMapper Mapper { get; set; }
+        [Inject] private IConfiguration config { get; set; }
 
         private VariantModel variantModel = new();
         private Image image = new();
@@ -32,12 +34,13 @@ namespace Inventory.Shared
                 {
                     if (image.ImageData != null)
                         variantModel.Image = image;
+                    else
+                        variantModel.Image = null;
 
                     Product.Variants.Add(Mapper.Map<VariantEntity>(variantModel));
                     await ProductRepository.Update(Product);
                     CancelVariant();
                     GetVariants();
-
                 }
                 catch (Exception ex)
                 {
@@ -48,7 +51,40 @@ namespace Inventory.Shared
 
         public async Task EditVariant()
         {
-
+            if (variantModel != null)
+            {
+                try
+                {
+                    if (image.Id != Guid.Empty)
+                    {
+                        var img = await ImageRepository.GetById(variantModel.Image.Id);
+                        if(img != null){
+                            img.ImageTitle = image.ImageTitle;
+                            img.ImageData = image.ImageData;
+                            variantModel.Image = img;
+                        }
+                    } else if (image.ImageData != null)
+                    {
+                        variantModel.Image = image;
+                    }
+                        
+                    var variant = Product.Variants.FirstOrDefault(v => v.Id == variantModel.Id);
+                    if (variant != null)
+                    {
+                        variant.VariantId = variantModel.VariantId;
+                        variant.Name = variantModel.Name;
+                        variant.StockInHand = variantModel.StockInHand;
+                        variant.Image = variantModel.Image;
+                        await ProductRepository.Update(Product);
+                        CancelVariant();
+                        GetVariants();
+                    }
+                }
+                catch (Exception ex)
+                {
+                    Logger.LogError("Create variant: " + ex.Message);
+                }
+            }
         }
 
         public void CancelVariant()
@@ -68,30 +104,66 @@ namespace Inventory.Shared
             try
             {
                 var variant = await VariantRepository.GetById(id);
-                await VariantRepository.Delete(variant);
-                GetVariants();
+                if (variant != null)
+                {
+                    await DeleteFile(variant.Image);
+                    await VariantRepository.Delete(variant);
+                    GetVariants();
+                }
             }
             catch (Exception ex)
             {
                 Logger.LogError("Delete variant error: " + ex.Message);
             }
-           
+
         }
 
         public async Task UploadFile(InputFileChangeEventArgs eventArgs)
         {
-            image.ImageTitle = eventArgs.File.Name;
-            using Stream stream = eventArgs.File.OpenReadStream();
-            using MemoryStream ms = new MemoryStream();
-            await stream.CopyToAsync(ms);
-            image.ImageData = ms.ToArray();
+            try
+            {
+                if (eventArgs.File.ContentType == "image/jpeg")
+                {
+                    var resizedFile = await eventArgs.File.RequestImageFileAsync(eventArgs.File.ContentType, 640, 480);
+                    image.ImageTitle = resizedFile.Name;
+                    var maxSizeUploadedFile = int.Parse(config.GetSection("FileMaxSize").Value);
+                    using Stream stream = resizedFile.OpenReadStream(maxSizeUploadedFile);
+                    using MemoryStream ms = new MemoryStream();
+                    await stream.CopyToAsync(ms);
+                    image.ImageData = ms.ToArray();
+                }
+            }
+            catch (Exception ex)
+            {
+                Logger.LogError("Upload file error: " + ex.Message);
+            }
+        }
 
+        public async Task DeleteFile(Image image)
+        {
+            if (image != null)
+            {
+                try
+                {
+                    await ImageRepository.Delete(image);
+                    GetVariants();
+                }
+                catch (Exception ex)
+                {
+                    Logger.LogError("Delete file error: " + ex.Message);
+                }
+            }
         }
 
         public void GetVariants()
         {
-            if (Product != null && Product.Variants.Count != 0)
+            if (Product != null)
                 variants = Product.Variants.Select(v => Mapper.Map<VariantModel>(v)).ToList();
+        }
+
+        public string GetImage(Image image)
+        {
+            return $"data:image/jpg;base64,{Convert.ToBase64String(image.ImageData)}";
         }
     }
 }
